requirements.txt
amqp==2.2.2
beautifulsoup4==4.6.0
billiard==3.5.0.3
bleach==2.0.0
boto==2.48.0
boto3==1.5.22
botocore==1.8.36
branca==0.2.0
celery==4.1.0
celery-haystack==0.10
chardet==3.0.4
colorama==0.3.9
decorator==4.0.11
Django==1.10.8
django-appconf==1.0.2
django-contrib-comments==1.8.0
django-cors-headers==2.1.0
django-disqus==0.5
django-endless-pagination==2.0
django-haystack==2.7.0
django-modelcluster==3.1
django-multiselectfield==0.1.8
django-storages==1.6.5
django-taggit==0.22.2
django-treebeard==4.2.0
djangorestframework==3.6.4
docutils==0.14
elasticsearch==2.4.1
elasticsearch-dsl==6.1.0
entrypoints==0.2.2
folium==0.3.0
future==0.16.0
geopy==1.11.0
grappelli-safe==0.4.7
html5lib==0.999999999
ipaddress==1.0.19
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
jedi==0.10.2
Jinja2==2.9.6
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.1
jupyter-console==5.1.0
jupyter-core==4.3.0
kombu==4.1.0
lxml==3.8.0
MarkupSafe==1.0
Mezzanine==4.2.3
mistune==0.7.4
mock==2.0.0
nbconvert==5.2.1
nbformat==4.3.0
notebook==5.0.0
numpy==1.13.1
oauthlib==2.0.6
olefile==0.44
opencv-contrib-python==3.3.0.9
opencv-python==3.3.0
pandas==0.20.2
pandocfilters==1.4.1
pbr==3.1.1
pickleshare==0.7.4
Pillow==4.2.1
prompt-toolkit==1.0.14
pygame==1.9.3
Pygments==2.2.0
PyJWT==1.5.2
pynput==1.3.7
PySocks==1.6.7
python-dateutil==2.6.0
pytz==2017.2
pyzmq==16.0.2
qtconsole==4.3.0
requests==2.12.4
requests-oauthlib==0.8.0
s3transfer==0.1.12
simplegeneric==0.8.1
six==1.10.0
testpath==0.3.1
tornado==4.5.1
traitlets==4.3.2
twilio==6.5.2
tzlocal==1.5.1
Unidecode==1.0.22
urllib3==1.21.1
vine==1.1.4
virtualenv==15.1.0
wagtail==1.13.1
wagtail-condensedinlinepanel==0.4.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==2.0.0
Willow==1.0
wxPython==4.0.0b1
